+++ griptape/config/amazon_bedrock_structure_config.py
@@ -1,4 +1,4 @@
-from attrs import define, field, Factory
+from attrs import Factory, define, field
 
 from griptape.config import StructureConfig
 from griptape.drivers import (