+++ tests/unit/structures/test_pipeline.py
@@ -67,7 +67,7 @@ class TestPipeline:
     def test_with_no_task_memory(self):
         pipeline = Pipeline()
 
-        pipeline.add_task(ToolkitTask(tools=[MockTool(off_prompt=False)]))
+        pipeline.add_task(ToolkitTask(tools=[MockTool()]))
 
         assert isinstance(pipeline.tasks[0], ToolkitTask)
         assert pipeline.tasks[0].tools[0].input_memory is not None
@@ -90,7 +90,7 @@ class TestPipeline:
         embedding_driver = MockEmbeddingDriver()
         pipeline = Pipeline(embedding_driver=embedding_driver)
 
-        pipeline.add_task(ToolkitTask(tools=[MockTool(off_prompt=False)]))
+        pipeline.add_task(ToolkitTask(tools=[MockTool()]))
 
         storage = list(pipeline.task_memory.artifact_storages.values())[0]
         assert isinstance(storage, TextArtifactStorage)
@@ -109,7 +109,7 @@ class TestPipeline:
     def test_without_task_memory(self):
         pipeline = Pipeline(task_memory=None)
 
-        pipeline.add_task(ToolkitTask(tools=[MockTool(off_prompt=False)]))
+        pipeline.add_task(ToolkitTask(tools=[MockTool()]))
 
         assert isinstance(pipeline.tasks[0], ToolkitTask)
         assert pipeline.tasks[0].tools[0].input_memory is None