+++ tests/unit/tasks/test_tool_task.py
@@ -147,7 +147,7 @@ class TestToolTask:
         )
 
     def test_run_without_memory(self, agent):
-        task = ToolTask(tool=MockTool(off_prompt=False))
+        task = ToolTask(tool=MockTool())
 
         agent.add_task(task)
 
@@ -168,9 +168,9 @@ class TestToolTask:
 
         subtask.structure = agent
 
-        memory.process_output(MockTool(off_prompt=False).test, subtask, TextArtifact("foo"))
+        memory.process_output(MockTool().test, subtask, TextArtifact("foo"))
 
-        task = ToolTask(tool=MockTool(off_prompt=False))
+        task = ToolTask(tool=MockTool())
 
         agent.add_task(task)
 
@@ -179,7 +179,7 @@ class TestToolTask:
         assert "You have access to additional contextual information" in system_template
 
     def test_actions_schema(self):
-        tool = MockTool(off_prompt=False)
+        tool = MockTool()
         task = ToolTask("test", tool=tool)
 
         Agent().add_task(task)