+++ tests/unit/tools/test_outpainting_image_variation_client.py
@@ -23,15 +23,11 @@ class TestOutpaintingImageGenerationClient:
 
     @pytest.fixture
     def image_generator(self, image_generation_engine, image_loader) -> OutpaintingImageGenerationClient:
-        return OutpaintingImageGenerationClient(
-            engine=image_generation_engine, image_loader=image_loader, off_prompt=False
-        )
+        return OutpaintingImageGenerationClient(engine=image_generation_engine, image_loader=image_loader)
 
     def test_validate_output_configs(self, image_generation_engine) -> None:
         with pytest.raises(ValueError):
-            OutpaintingImageGenerationClient(
-                engine=image_generation_engine, output_dir="test", output_file="test", off_prompt=False
-            )
+            OutpaintingImageGenerationClient(engine=image_generation_engine, output_dir="test", output_file="test")
 
     def test_image_outpainting(self, image_generator) -> None:
         image_generator.engine.run.return_value = Mock(
@@ -54,7 +50,7 @@ class TestOutpaintingImageGenerationClient:
     def test_image_outpainting_with_outfile(self, image_generation_engine, image_loader) -> None:
         outfile = f"{tempfile.gettempdir()}/{str(uuid.uuid4())}.png"
         image_generator = OutpaintingImageGenerationClient(
-            engine=image_generation_engine, output_file=outfile, image_loader=image_loader, off_prompt=False
+            engine=image_generation_engine, output_file=outfile, image_loader=image_loader
         )
 
         image_generator.engine.run.return_value = Mock(  # pyright: ignore