+++ tests/unit/tools/test_sql_client.py
@@ -20,9 +20,7 @@ class TestSqlClient:
 
     def test_execute_query(self, driver):
         with sqlite3.connect(":memory:"):
-            client = SqlClient(
-                sql_loader=SqlLoader(sql_driver=driver), table_name="test_table", engine_name="sqlite", off_prompt=False
-            )
+            client = SqlClient(sql_loader=SqlLoader(sql_driver=driver), table_name="test_table", engine_name="sqlite")
             result = client.execute_query({"values": {"sql_query": "SELECT * from test_table;"}})
 
             assert len(result.value) == 1
@@ -34,7 +32,6 @@ class TestSqlClient:
             table_name="test_table",
             table_description="foobar",
             engine_name="sqlite",
-            off_prompt=False,
         )
         description = client.activity_description(client.execute_query)