+++ tests/unit/tools/test_variation_image_generation_client.py
@@ -23,18 +23,12 @@ class TestVariationImageGenerationClient:
 
     @pytest.fixture
     def image_generator(self, image_generation_engine, image_loader) -> VariationImageGenerationClient:
-        return VariationImageGenerationClient(
-            engine=image_generation_engine, image_loader=image_loader, off_prompt=False
-        )
+        return VariationImageGenerationClient(engine=image_generation_engine, image_loader=image_loader)
 
     def test_validate_output_configs(self, image_generation_engine, image_loader) -> None:
         with pytest.raises(ValueError):
             VariationImageGenerationClient(
-                engine=image_generation_engine,
-                output_dir="test",
-                output_file="test",
-                image_loader=image_loader,
-                off_prompt=False,
+                engine=image_generation_engine, output_dir="test", output_file="test", image_loader=image_loader
             )
 
     def test_image_variation(self, image_generator) -> None:
@@ -57,7 +51,7 @@ class TestVariationImageGenerationClient:
     def test_image_variation_with_outfile(self, image_generation_engine, image_loader) -> None:
         outfile = f"{tempfile.gettempdir()}/{str(uuid.uuid4())}.png"
         image_generator = VariationImageGenerationClient(
-            engine=image_generation_engine, output_file=outfile, image_loader=image_loader, off_prompt=False
+            engine=image_generation_engine, output_file=outfile, image_loader=image_loader
         )
 
         image_generator.engine.run.return_value = Mock(  # pyright: ignore